Export projects from index directly
